tests: move sendActions from gristUtils to gristWebDriverUtils
Part of refactor to sync with grist-widget.
