+++ test/nbrowser/gristWebDriverUtils.ts
@@ -8,6 +8,7 @@
  * easily.
  */
 
+import { DocAction, UserAction } from 'app/common/DocActions';
 import { WebDriver, WebElement } from 'mocha-webdriver';
 
 type SectionTypes = 'Table'|'Card'|'Card List'|'Chart'|'Custom'|'Form';
@@ -256,6 +257,32 @@ export class GristWebDriverUtils {
   public async reloadDoc() {
     await this.driver.navigate().refresh();
     await this.waitForDocToLoad();
+  }
+
+  /**
+   * Sends UserActions using client api from the browser.
+   */
+  public async sendActions(actions: (DocAction | UserAction)[], optTimeout: number = 2000) {
+    await this.driver.manage().setTimeouts({
+      script: optTimeout, /* milliseconds */
+    });
+
+    // Make quick test that we have a list of actions not just a single action, by checking
+    // if the first element is an array.
+    if (actions.length && !Array.isArray(actions[0])) {
+      throw new Error('actions argument should be a list of actions, not a single action');
+    }
+
+    const result = await this.driver.executeAsyncScript(`
+    const done = arguments[arguments.length - 1];
+    const prom = gristDocPageModel.gristDoc.get().docModel.docData.sendActions(${JSON.stringify(actions)});
+    prom.then(() => done(null));
+    prom.catch((err) => done(String(err?.message || err)));
+  `);
+    if (result) {
+      throw new Error(result as string);
+    }
+    await this.waitForServer();
   }
 }
